Responding to review comments.
